extract manifest structs
